Limit workspace numbers within 0..INT32_MAX
Before this commit, large workspace numbers treated oddly:

   $ i3-msg 'rename workspace to 1234567890'
   # displayed in i3bar as `0`

   $ i3-msg 'rename workspace to 4294967200'
   $ i3-msg -t get_workspaces | jq '.[]|select(.focused).num'
   -96 # int32_t overflow

   $ i3-msg 'rename workspace to 99999999999999999999'
   $ i3-msg -t get_workspaces | jq '.[]|select(.focused).num'
   -1 # treated as unnumbered

This commit puts a consistent limit on workspace numbers.  Now
workspaces with numbers beyond INT32_MAX are treated as unnumbered.

+++ include/util.h
@@ -86,7 +86,7 @@ bool layout_from_name(const char *layout_str, layout_t *out);
  * interpreted as a "named workspace".
  *
  */
-long ws_name_to_number(const char *name);
+int ws_name_to_number(const char *name);
 
 /**
  * Updates *destination with new_value and returns true if it was changed or false
